backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.6,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
pymysql>=1.1,<2.0
redis>=5.2,<6.0
celery>=5.4,<6.0
elasticsearch>=8.15,<9.0
minio>=7.2,<8.0
python-multipart>=0.0.17,<1.0
email-validator>=2.2,<3.0
pytest>=8.3,<9.0
httpx>=0.28,<1.0
pypdf>=5.4,<6.0
python-docx>=1.1,<2.0
